# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is cxy api docs.",
        "title": "用户聊天系统",
        "contact": {
            "name": "go-swagger帮助文档"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "paths": {
        "/index": {
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/create-user": {
            "get": {
                "summary": "注册用户",
                "parameters": [
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "用户名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query"
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "密码",
                        "name": "passwordv2",
                        "in": "query"
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"userList\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/delete-user": {
            "delete": {
                "summary": "删除用户",
                "parameters": [
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"userList\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/get-all-user": {
            "get": {
                "summary": "获取所有用户",
                "responses": {
                    "200": {
                        "description": "code\",\"userList\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/get-user-by-name": {
            "get": {
                "summary": "根据用户名获取用户",
                "parameters": [
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "用户名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/update-user": {
            "put": {
                "summary": "修改用户",
                "parameters": [
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "用户id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "用户名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "type": "string",
                        "description": "用户手机号",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"userList\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}